requirements.txt
# Cloud SDKs
boto3==1.34.0
azure-storage-blob==12.19.0
google-cloud-storage==2.13.0

# CLI and utilities
click==8.1.7
python-dotenv==1.0.0
colorama==0.4.6
rich==13.7.0

# Logging and monitoring
structlog==23.2.0

# Security and validation
cryptography==41.0.8
pydantic==2.5.0

# Retry and resilience
tenacity==8.2.3

# Testing (optional)
pytest==7.4.4
pytest-mock==3.12.0
